+++ src/components/Carousel/Carousel.module.scss
@@ -1,15 +1,17 @@
-$color-green: #36d6ae;
 $color-white: #fff;
 $color-shadow-lamp-green: rgba(54, 214, 174, 0.5);
 $color-shadow-lamp-white: rgba(255, 255, 255, 0.44);
 $color-border-lamp-green: #00edeb;
 $color-border-lamp-white: #e2e2e2;
 
+$color-shadow-lamp-blue: rgba(36, 189, 219, 0.705);
+$color-border-lamp-blue: var(--blue-light);
+
 @mixin linear-gradient-gray-el {
   content: '';
   height: 100%;
   position: absolute;
-  width: 20%;
+  width: 10%;
   bottom: 0;
   opacity: 11%;
 }
@@ -43,7 +45,7 @@ $color-border-lamp-white: #e2e2e2;
 
 @mixin light-lamp($color-bgr, $color-border, $color-shadow) {
   background: $color-bgr;
-  border: 1px solid $color-border;
+  border: 0.5px solid $color-border;
   box-sizing: border-box;
   box-shadow: 0px 0px 5px 1px $color-shadow;
 }
@@ -80,14 +82,12 @@ $color-border-lamp-white: #e2e2e2;
     display: flex;
 
     .slide {
-      // background: #dadada;
       display: flex;
       flex-direction: column;
       align-items: center;
       justify-content: center;
       position: relative;
-      cursor: pointer;
-      color: #36d6ae;
+      color: var(--primary-color);
 
       &::before {
         @include linear-gradient-gray-el;
@@ -117,9 +117,6 @@ $color-border-lamp-white: #e2e2e2;
     }
 
     .slide.left {
-      cursor: pointer;
-      color: $color-green;
-      //  margin-left: 10px;
       &::before {
         @include init-pseudo-el-gren-default;
         @include linear-gradient-green(90deg);
@@ -138,8 +135,6 @@ $color-border-lamp-white: #e2e2e2;
       }
     }
     .slide.right {
-      // margin-right: 10px;
-      cursor: pointer;
       &::before {
         @include init-pseudo-el-gren-default;
         @include linear-gradient-green(270deg);
@@ -155,7 +150,6 @@ $color-border-lamp-white: #e2e2e2;
         position: absolute;
         height: 100%;
       }
-      color: $color-green;
     }
 
     .containerContent {
@@ -168,31 +162,26 @@ $color-border-lamp-white: #e2e2e2;
     }
 
     .lamp.active {
-      &::before {
-        content: '';
-        position: absolute;
-        top: 0;
-        bottom: 0;
-        left: 0;
-        @include light-lamp(
-          $color-white,
-          $color-border-lamp-green,
-          $color-shadow-lamp-green
-        );
-        border-radius: 0px 1px 1px 0px;
-      }
-
+      &::before,
       &::after {
         content: '';
         position: absolute;
         top: 0;
         bottom: 0;
-        right: 0;
+
         @include light-lamp(
           $color-white,
           $color-border-lamp-green,
           $color-shadow-lamp-green
         );
+      }
+      &::before {
+        left: 0;
+        border-radius: 0px 1px 1px 0px;
+      }
+
+      &::after {
+        right: 0;
         border-radius: 1px 0px 0px 1px;
       }
     }
@@ -258,4 +247,47 @@ $color-border-lamp-white: #e2e2e2;
 }
 .transition {
   transition: ease all 0.5s;
+}
+
+.color_ {
+  &blue {
+    .slide {
+      color: var(--blue-light) !important;
+    }
+
+    .slides {
+      .lamp {
+        &.active {
+          &::before,
+          &::after {
+            @include light-lamp(
+              $color-white,
+              $color-border-lamp-blue,
+              $color-shadow-lamp-blue
+            );
+          }
+        }
+
+        &.left {
+          &::after {
+            @include light-lamp(
+              $color-white,
+              $color-border-lamp-blue,
+              $color-shadow-lamp-blue
+            );
+          }
+        }
+
+        &.right {
+          &::before {
+            @include light-lamp(
+              $color-white,
+              $color-border-lamp-blue,
+              $color-shadow-lamp-blue
+            );
+          }
+        }
+      }
+    }
+  }
 }
