+++ src/containers/portal/citizenship/index.tsx
@@ -32,7 +32,7 @@ import { CYBER } from '../../../utils/config';
 import useSetActiveAddress from '../../../hooks/useSetActiveAddress';
 import { steps } from './utils';
 import Info from './Info';
-import Carousel from '../gift/carousel1/Carousel';
+import Carousel from '../../../components/Carousel/CarouselOld/CarouselOld';
 import { useAdviser } from 'src/features/adviser/context';
 import { useBackend } from 'src/contexts/backend';
