--- src/containers/temple/components/index.js
@@ -1,4 +1,0 @@
-import Canvas from './canvasOne';
-import Carousel from './Carousel/Carousel';
-
-export { Canvas, Carousel };
