+++ src/containers/temple/index.jsx
@@ -1,11 +1,12 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { MainContainer } from '../portal/components';
-import { Carousel, Canvas } from './components';
 import { BOOT_ICON } from '../portal/utils';
 import { PlayContent, PlayBanerContent as PlayBannerContent } from './pages';
 import { ActionBar, ContainerGradientText } from '../../components';
 import styles from './styles.module.scss';
+import Carousel from 'components/Carousel/Carousel';
+import Canvas from './components/canvasOne';
 
 const itemCarousel = [
   { title: 'compute' },
